requirements.txt
azure-common==1.1.28
azure-core==1.34.0
azure-identity==1.23.0
azure-mgmt-core==1.5.0
azure-mgmt-monitor==6.0.2
boto3==1.38.39
botocore==1.38.39
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.4
deepl==1.22.0
filelock==3.18.0
fsspec==2025.5.1
huggingface-hub==0.33.0
idna==3.10
isodate==0.7.2
jmespath==1.0.1
lxml==5.4.0
msal==1.32.3
msal-extensions==1.3.1
numpy==2.3.0
packaging==25.0
pillow==11.2.1
pycparser==2.22
PyJWT==2.10.1
PyMuPDF==1.26.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
s3transfer==0.13.0
safetensors==0.5.3
setuptools==80.9.0
six==1.17.0
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.52.4
typing_extensions==4.14.0
urllib3==2.4.0
wheel==0.45.1
